backend/pyproject.toml
[project]
name = "telegramemojis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
    "python-magic",
    "matplotlib",
    "numpy",
    "requests>=2.32.3",
]
